Reorganised goals and social files
