+++ lib/main.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'themecolors.dart';
 import 'auth.dart';
 import 'signinpage.dart';
-import 'social.dart';
+import 'package:pudding_flutter/social/social.dart';
 import 'timetable.dart';
-import 'goals.dart';
+import 'package:pudding_flutter/goals/goals.dart';
 import 'dashboard.dart';
 
 /// Set this to false if the initial sign in page is creating issues for you!
